fix bug

+++ run.sh
@@ -2,7 +2,7 @@
 #作者: entr0pia (风沐白)
 #文件: run.sh
 #描述: 移除Android固件(如Sony Xpeira 5II)的boot.img中的system_ext分区, 修复patch magisk后的bootloop问题
-#版本: v2.0
+#版本: v2.0.1
 
 img=$1
 workspace=$(pwd)
@@ -22,7 +22,13 @@ function patchMagisk(){
     mv ../image-new.img .
     git fetch --all && git reset --hard origin/master
     ./boot_patch.sh image-new.img
-    mv new-boot.img ../image-new.img
+    if [ -s "new-boot.img" ]; then
+        mv new-boot.img ../image-new.img
+    else
+        echo -e "\033[37;41m 自动 Patch 失败, 请手动刷入 Magisk \033[0m"
+        echo -e "\033[37;41m Auto Patch Failed, Need to Flash Magisk Manually \033[0m"
+        mv image-new.img ..
+    fi
     git clean -xdf
 }
